Move create_reserved_ip & DEFAULT_REGIONS to general conftest & helpers
