+++ tests/integration/linodes/helpers.py
@@ -1,18 +1,13 @@
 import json
-import os
 import time
 
 from tests.integration.helpers import (
     BASE_CMDS,
+    DEFAULT_REGION,
     exec_test_command,
 )
 
 DEFAULT_RANDOM_PASS = exec_test_command(["openssl", "rand", "-base64", "32"])
-DEFAULT_REGION = (
-    "pl-labkrk-2"
-    if "devcloud" in os.getenv("LINODE_CLI_API_HOST", "")
-    else "us-ord"
-)
 
 DEFAULT_TEST_IMAGE = exec_test_command(
     [
