+++ tests/integration/networking/fixtures.py
@@ -1,37 +1,10 @@
-import json
-
 import pytest
 
-from tests.integration.helpers import (
-    BASE_CMDS,
-    delete_target_id,
-    exec_test_command,
-)
+from tests.integration.helpers import delete_target_id
 from tests.integration.linodes.helpers import (
-    DEFAULT_REGION,
     create_linode,
     create_linode_and_wait,
 )
-
-
-@pytest.fixture
-def create_reserved_ip(request):
-    tags = getattr(request, "param", None)
-    command = BASE_CMDS["networking"] + [
-        "reserved-ip-add",
-        "--region",
-        DEFAULT_REGION,
-        "--json",
-    ]
-
-    if tags:
-        command += ["--tags", tags]
-
-    result = json.loads(exec_test_command(command))[0]
-
-    yield result
-
-    delete_target_id("networking", result["address"], "reserved-ip-delete")
 
 
 @pytest.fixture(scope="package")
@@ -60,11 +33,3 @@ def get_linode_ids_shared_ipv4(linode_cloud_firewall):
 
     for id_num in linode_ids:
         delete_target_id(target="linodes", id=id_num)
-
-
-def get_command_heads_and_vals(command):
-    result = exec_test_command(command).splitlines()
-    headers = [item for item in result[0].split(",")]
-    values = [item for item in result[1].split(",")]
-
-    return headers, values
